Created MessageService.java
Moved actions from the BroadcastReceiver to a Service.
TODO: it doesn't work because I moved things into a bundle
TODO: it seems incorrect that an IntentService have the location
listener. We'll need to consider what's the best way for the
activities to get this information since it only really matters
when they're available...